requirements.txt
# python 3.12.8
librosa==0.9.0
praat-parselmouth==0.4.3
pandas==2.2.0
setuptools==75.8.0
matplotlib==3.10.0
jupyterlab==4.3.4